+++ packages/soda-antd/src/index.ts
@@ -1,3 +1,4 @@
 export * from "@/components/NiceSearch"
 export * from "@/components/Theme"
 export * from "@/hooks/useScrollY"
+export * from "@/types/index"