+++ packages/soda-type/src/index.ts
@@ -1,6 +1,3 @@
-import { ReactNode } from "react"
-import { ColumnsType } from "antd/es/table"
-
 /** 判断两个类型是否全等 */
 export type Equal<X, Y> = (<T>() => T extends X ? 1 : 2) extends <T>() => T extends Y ? 1 : 2 ? true : false
 
@@ -118,26 +115,6 @@ export type CanJSONStringify<T> = _CanJSONStringify<T>
 
 /** 获取字符串提示 */
 export type TipString<T extends string> = T | (string & {})
-
-type _Column<T, K extends keyof T = keyof T> = Omit<ColumnsType<T>[0], "dataIndex" | "render"> &
-    (
-        | (K extends keyof T
-              ? {
-                    dataIndex: K
-                } & (T[K] extends ReactNode
-                    ? { render?: (value: T[K], record: T, index: number) => ReactNode }
-                    : { render: (value: T[K], record: T, index: number) => ReactNode })
-              : never)
-        | ({
-              dataIndex?: undefined
-          } & (T extends ReactNode
-              ? { render?: (value: T, record: T, index: number) => ReactNode }
-              : { render: (value: T, record: T, index: number) => ReactNode }))
-    )
-
-export type Column<T> = _Column<T>
-
-export type Columns<T> = Column<T>[]
 
 /** 编辑器 props */
 export interface EditorProps<Data, Error = unknown> {