+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # deepsea-components
 
+## 5.9.7
+
+### Patch Changes
+
+- 将 Column 类型移动到 soda-antd 包中，升级 react 相关依赖
+
 ## 5.9.6
 
 ### Patch Changes